FOR 1.9 SUPPORT Fix small shifts (use double) (cherry picked from commit 97a0752aae462dab87f2b7a67e6f3d0d0cba6390)
Conflicts:

	native/src/common2.h
	native/src/java_wrap.cpp
	native/src/rendering.cpp

+++ native/src/rendering.cpp
@@ -27,8 +27,8 @@ void calcPoint(std::pair<int, int> const & c, RenderingContext & rc)
 {
     rc.pointCount++;
 
-	float tx = c.first/ (rc.tileDivisor);
-	float ty = c.second / (rc.tileDivisor);
+	double tx = c.first/ (rc.tileDivisor);
+	double ty = c.second / (rc.tileDivisor);
 
     float dTileX = tx - rc.getLeft();
     float dTileY = ty - rc.getTop();
